# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: FitAI API
  description: |
    API completa para FitAI - Aplicación de fitness con IA para el mercado chileno.
    
    ## Autenticación
    La mayoría de endpoints requieren autenticación mediante Clerk. Include el token en el header:
    ```
    Authorization: Bearer YOUR_CLERK_TOKEN
    ```
    
    ## Planes de Suscripción
    - **Free**: Funcionalidades básicas
    - **Premium**: IA avanzada, analytics, social features ($7,990 CLP/mes)
    - **Pro**: Todas las funcionalidades premium + análisis avanzado ($14,990 CLP/mes)
    
    ## Rate Limits
    - API general: 1000 requests/hour
    - IA features:
      - Free: 3 rutinas, 10 consejos por día
      - Premium: 15 rutinas, 50 consejos por día  
      - Pro: 50 rutinas, 200 consejos por día
  version: 1.0.0
  contact:
    name: FitAI Support
    email: support@fitai.cl
  license:
    name: Proprietary
servers:
  - url: https://api.getfitia.com
    description: Production server
  - url: https://fitai-api.sweetspot-627.workers.dev
    description: Development server

# Security schemes
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk authentication token
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "ValidationError"
        message:
          type: string
          example: "Invalid input data"
        path:
          type: string
          example: "/api/v1/users/me"
    PaginationResponse:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
    
    # User schemas
...
    Exercise:
      type: object
      properties:
        id:
          type: string
          example: "ex_123"
        name:
          type: string
          example: "Bench Press"
        nameEs:
          type: string
          example: "Press de Banca"
        category:
          type: string
          example: "strength"
        muscleGroups:
          type: array
          items:
            type: string
          example: ["chest", "triceps"]
        equipment:
          type: string
          example: "barbell"
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
          example: "intermediate"
        instructions:
          type: array
          items:
            type: string
        instructionsEs:
          type: array
          items:
            type: string
        tips:
          type: array
          items:
            type: string
        videoUrl:
          type: string
          format: uri
        imageUrls:
          type: array
          items:
            type: string
            format: uri
    
    # Workout schemas
...
security:
  - ClerkAuth: []
paths:
...
  /api/v1/exercises:
    get:
      tags:
        - Exercises
      summary: Get All Exercises
      description: Obtiene el catálogo completo de ejercicios (público)
      security: []
      parameters:
        - name: category
          in: query
          description: Filtrar por categoría
          schema:
            type: string
        - name: muscle_group
          in: query
          description: Filtrar por grupo muscular
          schema:
            type: string
        - name: equipment
          in: query
          description: Filtrar por equipamiento
          schema:
            type: string
        - name: difficulty
          in: query
          description: Filtrar por dificultad
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
        - name: limit
          in: query
          description: Número máximo de resultados
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          description: Número de resultados a omitir
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista de ejercicios obtenida exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Exercise'
                      pagination:
                        $ref: '#/components/schemas/PaginationResponse'
  /api/v1/exercises/{id}:
    get:
      tags:
        - Exercises
      summary: Get Exercise by ID
      description: Obtiene un ejercicio específico por su ID
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID del ejercicio
          schema:
            type: string
      responses:
        '200':
          description: Ejercicio obtenido exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Exercise'
        '404':
          description: Ejercicio no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # AI endpoints
...